fix #180

+++ src/python_vali/src/PySurface.cpp
@@ -501,7 +501,7 @@ void Init_PySurface(py::module& m) {
                 cai.m_ptr = (CUdeviceptr)tup[0].cast<size_t>();
                 cai.m_read_only = tup[1].cast<bool>();
               } else if ("stream" == key) {
-                cai.m_stream = (CUstream)item.second.cast<int>();
+                cai.m_stream = (CUstream)item.second.cast<size_t>();
               } else if ("version" == key) {
                 auto const version = item.second.cast<int>();
                 if (version != 3) {
